# src/swagger_output.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "v1.0.0",
    "title": "Lend A Friend",
    "description": "Lend A Friend Mobile app REST API server"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": ""
    }
  ],
  "paths": {
...
    "/api/transactions/approve-transaction/{transactionId}": {
      "put": {
        "operationId": "TransactionsController.approveTransaction",
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "approvalDate": {
                    "$ref": "#/components/schemas/Date"
                  },
                  "userId": {
                    "type": "number"
                  }
                },
                "required": [],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Approve transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/approve-witness-transaction/{transactionId}": {
      "put": {
        "operationId": "TransactionsController.approveWitnessTransaction",
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "approvalDate": {
                    "$ref": "#/components/schemas/Date"
                  },
                  "userId": {
                    "type": "number"
                  }
                },
                "required": [],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Approve witness transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/change-transaction/{userId}": {
      "patch": {
        "operationId": "TransactionsController.changeTransaction",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Change transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/confirm-settlement/{transactionId}": {
      "patch": {
        "operationId": "TransactionsController.confirmSettlement",
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "userId": {
                    "type": "number"
                  }
                },
                "required": [],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Confirm settlement",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{userId}": {
      "get": {
        "operationId": "TransactionsController.getTransactions",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Get transactions",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/initiate-transaction": {
      "post": {
        "operationId": "TransactionsController.initiateTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Initiate transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/rate-transaction/{userId}": {
      "patch": {
        "operationId": "TransactionsController.rateTransaction",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Rate transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/refuse-transaction/{transactionId}": {
      "patch": {
        "operationId": "TransactionsController.refuseTransaction",
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "userId": {
                    "type": "number"
                  }
                },
                "required": [],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Refuse transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/reject-settlement/{transactionId}": {
      "patch": {
        "operationId": "TransactionsController.rejectSettlement",
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "userId": {
                    "type": "number"
                  }
                },
                "required": [],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Reject settlement",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/request-settlement/{transactionId}/{userId}": {
      "get": {
        "operationId": "TransactionsController.requestSettlement",
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Request settlement",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/settle-transaction/{transactionId}": {
      "put": {
        "operationId": "TransactionsController.settleTransaction",
        "parameters": [
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "pattern": "[^\\/#\\?]+?",
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "settlementDate": {
                    "$ref": "#/components/schemas/Date"
                  },
                  "userId": {
                    "type": "number"
                  },
                  "receipt": {
                    "type": "string"
                  }
                },
                "required": [],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Settle transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}